requirements.txt
coverage
flask
psycopg2-binary
pytest